Extract slug spec from posts spec
System specs should cover 1 specific feature and not contain multiple
`it` blocks.

This follows guidance from Apply for Teacher Training's testing styleguide:

- https://github.com/DFE-Digital/apply-for-teacher-training/blob/main/docs/testing-styleguide.md
